requirements.txt
black~=23.10.1
telepot~=12.7
python-dotenv~=1.0.0